Moved state up and all of the handlers so that the state is shared between the two components
