+++ app/javascript/components/EligibilityDates.jsx
@@ -1,60 +1,18 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import { format, isWithinRange, eachDay } from 'date-fns';
+import { format } from 'date-fns';
 
 import PresenceDates from '../bundles/Presence/components/PresenceDates';
 
-const checkEligibility = (eligibleDateRanges, date = new Date()) => {
-  let eligible = false;
-
-  for (let index = 0; index < eligibleDateRanges.length; index++) {
-    const { start, end } = eligibleDateRanges[index];
-    if (isWithinRange(date, start, end)) {
-      eligible = true;
-      break;
-    }
-  }
-
-  return eligible;
-};
-
 export default class EligibilityDates extends React.Component {
-  state = {
-    selectedDate: new Date(),
-    isEligible: checkEligibility(this.props.eligibleDateRanges)
-  };
-
-  static propTypes = {
-    eligibleDateRanges: PropTypes.array.isRequired
-  };
-
-  allDaysInRange = () => {
-    const { eligibleDateRanges } = this.props;
-    const allDaysInRange = eligibleDateRanges.map(({ start, end }) => {
-      return eachDay(start, end, 1);
-    });
-    return allDaysInRange.reduce((acc, val) => acc.concat(val));
-  };
-
-  onDateChange = (date) => {
-    const { eligibleDateRanges } = this.props;
-    let newDate = new Date(date);
-
-    this.setState({
-      selectedDate: newDate,
-      isEligible: checkEligibility(eligibleDateRanges, newDate)
-    });
-  };
-
   render() {
-    const { selectedDate, isEligible } = this.state;
-    const { eligibleDateRanges } = this.props;
-
-    const highlightWithRanges = [
-      {
-        'is-within-range': this.allDaysInRange()
-      }
-    ];
+    const {
+      eligibleDateRanges,
+      selectedDate,
+      isEligible,
+      onDateChange,
+      highlightWithRanges
+    } = this.props;
 
     return (
       <div className="results">
@@ -63,7 +21,7 @@ export default class EligibilityDates extends React.Component {
           <PresenceDates
             isEligible={isEligible}
             selectedDate={selectedDate}
-            onDateChange={this.onDateChange}
+            onDateChange={onDateChange}
             highlightDates={highlightWithRanges}
           />
