+++ app/views/clients/show.html.erb
@@ -1,14 +1,13 @@
 <p id="notice"><%= notice %></p>
 
 <%= react_component("ShowClient", props: {
-  id: @client.id,
+  clientId: @client.id,
   identities: @client.identities.as_json(methods: [:country_of_birth_name, :issuing_state_name]),
   eligibleDateRanges:[
     {start:"Dec 17 2018", end:"Jan 13 2019"},
     {start:"Feb 17 2019", end:"April 13 2019"},
     {start:"Jun 11 2019", end:"Jul 3 2019"},
   ],
-  isEligible: true,
   totalDays: 1512,
   years: [
     {
